+++ api/src/resources/user.rs
@@ -1,7 +1,9 @@
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
+use serde_with::{DeserializeFromStr, SerializeDisplay};
 use std::{
     collections::{HashMap, HashSet},
+    fmt::Display,
     str::FromStr,
 };
 
@@ -229,40 +231,75 @@ impl FromStr for ProjectPermission {
     }
 }
 
-#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq, Hash)]
-#[serde(untagged)]
+#[derive(Debug, Clone, DeserializeFromStr, SerializeDisplay, PartialEq, Eq, Hash)]
 pub enum GlobalPermission {
-    // TODO(jcalero)[RE-978] There is a bug with the implementation of this enum that causes
-    // deserialization of non-Unknown properties to fail. See
-    // [RE-978](https://reinfer.atlassian.net/browse/RE-978) for more info.
-    #[serde(rename = "root")]
     Root,
-
-    #[serde(rename = "debug")]
     Debug,
-
-    #[serde(rename = "demo")]
     Demo,
-
-    #[serde(rename = "subscriptions-read")]
     SubscriptionsRead,
-
-    #[serde(rename = "artefacts-read")]
     ArtefactsRead,
-
-    #[serde(rename = "dialog")]
     LegacyDialog,
-
+    SupportTenantAdmin,
+    SupportUsersWrite,
+    DeploymentQuotaWrite,
+    TenantAdmin,
+    TenantQuotaWrite,
     Unknown(Box<str>),
 }
+
+const ROOT_AS_STR: &str = "root";
+const DEBUG_AS_STR: &str = "debug";
+const DEMO_AS_STR: &str = "demo";
+const SUBSCRIPTIONS_READ_AS_STR: &str = "subscriptions-read";
+const ARTEFACTS_READ_AS_STR: &str = "artefacts-read";
+const LEGACY_DIALOG_AS_STR: &str = "dialog";
+const SUPPORT_TENANT_ADMIN_AS_STR: &str = "support-tenant-admin";
+const SUPPORT_USERS_WRITE_AS_STR: &str = "support-users-write";
+const DEPLOYMENT_QUOTA_WRITE_AS_STR: &str = "deployment-quota-write";
+const TENANT_ADMIN_AS_STR: &str = "tenant-admin";
+const TENANT_QUOTA_WRITE_AS_STR: &str = "tenant-quota-write";
 
 impl FromStr for GlobalPermission {
     type Err = Error;
 
     fn from_str(string: &str) -> Result<Self> {
-        serde_json::de::from_str(&format!("\"{string}\"")).map_err(|_| Error::BadGlobalPermission {
-            permission: string.into(),
+        Ok(match string {
+            ROOT_AS_STR => GlobalPermission::Root,
+            DEBUG_AS_STR => GlobalPermission::Debug,
+            DEMO_AS_STR => GlobalPermission::Demo,
+            SUBSCRIPTIONS_READ_AS_STR => GlobalPermission::SubscriptionsRead,
+            ARTEFACTS_READ_AS_STR => GlobalPermission::ArtefactsRead,
+            LEGACY_DIALOG_AS_STR => GlobalPermission::LegacyDialog,
+            SUPPORT_TENANT_ADMIN_AS_STR => GlobalPermission::SupportTenantAdmin,
+            SUPPORT_USERS_WRITE_AS_STR => GlobalPermission::SupportUsersWrite,
+            TENANT_ADMIN_AS_STR => GlobalPermission::TenantAdmin,
+            TENANT_QUOTA_WRITE_AS_STR => GlobalPermission::TenantQuotaWrite,
+            DEPLOYMENT_QUOTA_WRITE_AS_STR => GlobalPermission::DeploymentQuotaWrite,
+            value => GlobalPermission::Unknown(value.into()),
         })
+    }
+}
+
+impl Display for GlobalPermission {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(
+            f,
+            "{}",
+            match self {
+                GlobalPermission::Root => ROOT_AS_STR,
+                GlobalPermission::Debug => DEBUG_AS_STR,
+                GlobalPermission::Demo => DEMO_AS_STR,
+                GlobalPermission::SubscriptionsRead => SUBSCRIPTIONS_READ_AS_STR,
+                GlobalPermission::ArtefactsRead => ARTEFACTS_READ_AS_STR,
+                GlobalPermission::LegacyDialog => LEGACY_DIALOG_AS_STR,
+                GlobalPermission::SupportTenantAdmin => SUPPORT_TENANT_ADMIN_AS_STR,
+                GlobalPermission::SupportUsersWrite => SUPPORT_USERS_WRITE_AS_STR,
+                GlobalPermission::TenantAdmin => TENANT_ADMIN_AS_STR,
+                GlobalPermission::TenantQuotaWrite => TENANT_QUOTA_WRITE_AS_STR,
+                GlobalPermission::DeploymentQuotaWrite => DEPLOYMENT_QUOTA_WRITE_AS_STR,
+                GlobalPermission::Unknown(value) => value.as_ref(),
+            }
+        )
     }
 }
 
@@ -287,8 +324,34 @@ pub struct PostUserResponse {
 
 #[cfg(test)]
 mod tests {
+
     use super::*;
 
+    #[test]
+    fn test_global_permissions_json_round_trip() {
+        let global_permissions = vec![
+            GlobalPermission::Root,
+            GlobalPermission::Debug,
+            GlobalPermission::Demo,
+            GlobalPermission::SubscriptionsRead,
+            GlobalPermission::ArtefactsRead,
+            GlobalPermission::LegacyDialog,
+            GlobalPermission::SupportTenantAdmin,
+            GlobalPermission::SupportUsersWrite,
+            GlobalPermission::TenantAdmin,
+            GlobalPermission::TenantQuotaWrite,
+            GlobalPermission::DeploymentQuotaWrite,
+            GlobalPermission::Unknown("new-perm".to_string().into_boxed_str()),
+        ];
+        let global_permissions_as_json_str = serde_json::to_string(&global_permissions).unwrap();
+
+        assert_eq!("[\"root\",\"debug\",\"demo\",\"subscriptions-read\",\"artefacts-read\",\"dialog\",\"support-tenant-admin\",\"support-users-write\",\"tenant-admin\",\"tenant-quota-write\",\"deployment-quota-write\",\"new-perm\"]", global_permissions_as_json_str);
+
+        let global_permissions_from_json_str: Vec<GlobalPermission> =
+            serde_json::from_str(&global_permissions_as_json_str).unwrap();
+
+        assert_eq!(global_permissions, global_permissions_from_json_str);
+    }
     #[test]
     fn unknown_project_permission_roundtrips() {
         let unknown_permission = ProjectPermission::from_str("unknown").unwrap();